Mudanca do input controlado
