+++ semana8/insta-lab-hooks/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import Post from './components/Post/Post';
 import styled from "styled-components"
 import SecaoNovaPostagem from './components/SecaoNovaPostagem/SecaoNovaPostagem';
@@ -10,66 +10,48 @@ const AppContainer = styled.div`
   align-items: center;
 `
 
-const App = () => {
+function App() {
 
   const [postsList, setPostsList] = useState([])
   const [post, setPost] = useState({})
 
-  const [inputUserName, setInputUserName] = useState('')
-  const [inputUserAvatar, setInputUserAvatar] = useState('')
-  const [inputLink, setInputLink] = useState('')
 
-  const handleInputUserName = (e) => {
-    setInputUserName(e.target.value)
-  }
-
-  const handleInputUserAvatar = (e) => {
-    setInputUserAvatar(e.target.value)
-  }
-
-  const handleInputLink = (e) => {
-    setInputLink(e.target.value)
-  }
-
-  const handlePostsList = () => {
-    setPost({
-      user: inputUserName,
-      avatar: inputUserAvatar,
-      link: inputLink
-    })
+  useEffect(() => {
     const newPostsList = postsList
     newPostsList.push(post)
     setPostsList(newPostsList)
+  },[post])
+
+  const handlePost = (name, avatar, link) => {
+    setPost({
+      user: name,
+      avatar: avatar,
+      link: link
+    })
   }
+
+  const allPosts = postsList.map((post) => {
+    return (
+      <Post
+        key={post.link}
+        nomeUsuario={post.user}
+        fotoUsuario={post.avatar}
+        fotoPost={post.link}
+      />
+    )
+  })
+
+  // console.log(post)
+  // console.log(postsList)
 
   return (
     <AppContainer>
       <SecaoNovaPostagem
-        inputUserName={inputUserName} onChangeName={handleInputUserName}
-        inputUserAvatar={inputUserAvatar} onChangeAvatar={handleInputUserAvatar}
-        inputLink={inputLink} onChangeLink={handleInputLink}
-        onClickCreate={handlePostsList}
+        onClickCreate={handlePost}
       />
-
-      {postsList.map((post) => {
-        return (
-          <Post
-            key={post.link}
-            nomeUsuario={post.user}
-            fotoUsuario={post.avatar}
-            fotoPost={post.link}
-          />
-        )
-      })}
-
-      {/* <Post
-        nomeUsuario={'Usuário'}
-        fotoUsuario={'https://picsum.photos/50/50'}
-        fotoPost={'https://picsum.photos/200/150'}
-      /> */}
+      {allPosts}
     </AppContainer>
   );
 }
-
 
 export default App;
