+++ semana8/insta-lab-hooks/src/components/SecaoNovaPostagem/SecaoNovaPostagem.js
@@ -1,28 +1,44 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-const SecaoNovaPostagem = (props) => {
+function SecaoNovaPostagem(props) {
+
+    const [inputUserName, setInputUserName] = useState('')
+    const [inputUserAvatar, setInputUserAvatar] = useState('')
+    const [inputLink, setInputLink] = useState('')
+
+    const handleInputUserName = (e) => {
+        setInputUserName(e.target.value)
+    }
+
+    const handleInputUserAvatar = (e) => {
+        setInputUserAvatar(e.target.value)
+    }
+
+    const handleInputLink = (e) => {
+        setInputLink(e.target.value)
+    }
 
     return (
         <div>
             <input
                 placeholder={"Nome de usuário"}
-                value={props.inputUser}
-                onChange={props.onChangeName}
+                value={inputUserName}
+                onChange={handleInputUserName}
 
             />
             <input
                 placeholder={"Imagem do avatar"}
-                value={props.inputAvatar}
-                onChange={props.onChangeAvatar}
+                value={inputUserAvatar}
+                onChange={handleInputUserAvatar}
 
             />
             <input
                 placeholder={"Link da imagem"}
-                value={props.inputLink}
-                onChange={props.onChangeLink}
+                value={inputLink}
+                onChange={handleInputLink}
 
             />
-            <button onClick={props.onClickCreate}>Postar</button>
+            <button onClick={() => props.onClickCreate(inputUserName, inputUserAvatar, inputLink)}>Postar</button>
         </div>
     )
 }
